fix: 데이터 타입 변경 및 훈련 데이터셋 사이즈 변경

+++ train/train.py
@@ -46,7 +46,7 @@ for epoch in range(epochs):
             torch.cuda.synchronize()
             for k, (x, padding_mask, y) in enumerate(loaded):
                 total_batch += 1
-                with autocast('cuda', torch.bfloat16):
+                with autocast('cuda', torch.float16):
                     pred = model(x, padding_mask)
                     loss = criterion(pred, y) / update_size
                 scaler.scale(loss).backward()
